requirements.txt
backcall==0.1.0
beautifulsoup4==4.6.0
boto3==1.7.52
botocore==1.10.52
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.6
django-filter==1.1.0
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
Faker==0.8.8
gunicorn==19.9.0
idna==2.7
inflect==0.3.1
ipython==6.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mixer==6.0.1
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
